% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_fig.R
\name{array_fig}
\alias{array_fig}
\title{Make figures from arrays}
\usage{
array_fig(array, module, xlimits, ylimits, background, cluster_results)
}
\arguments{
\item{array}{An array of all the pairwise Z-scores for a module}

\item{module}{the name of the KOs}

\item{xlimits}{a vector with the limits of the X axis}

\item{ylimits}{a vector with the limits of the Y axis}

\item{background}{The background distribution for individual KOs}

\item{cluster_results}{the results from the clustering algorithm}
}
\value{
a multi-panel figure. The first column is the distributions, each subsequent column is the cluster. Rows represent each individual KO
}
\description{
This function creates standardized figures to visualize differences between clusters
}
\examples{
for(i in 1:length(PHA_clustering_results_P_NRED$cl)) {
array_fig(PHA_module_P_NRED$combined,cluster_text_matrix[,i],c(-4,4),c(0,1),I_KOs_Background$random_pairwise_gene_correlation,PHA_clustering_results_P_NRED$cl)
}
}
