% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ave_Z_score_Func.R
\name{ave_Z_score_Func}
\alias{ave_Z_score_Func}
\title{Calculates the average Z Score}
\usage{
ave_Z_score_Func(combined_scores_array)
}
\arguments{
\item{combined_scores_array}{An array of all the pairwise Z-scores in a module}
}
\value{
a pairwise correlation matrix of average Z-scores
}
\description{
This function converts an array of Z-scores into a matrix of average scores.
(may be used on outputs for F11-F13)
}
\examples{
ave_Z_score_matrix <-ave_Z_score_Func(PHA_module_P_NRED$combined)
}
