% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Background_Distribution_Modules.R
\name{Background_Distribution_Modules}
\alias{Background_Distribution_Modules}
\title{Generate a background distibution of composite Z scores}
\usage{
Background_Distribution_Modules(N, Z)
}
\arguments{
\item{N}{The number of elements to be included in the randomly generated module}

\item{Z}{The number of iterations used to calculatea background distribution}
}
\value{
a list of vectors containing XXX & YYY
}
\description{
This function generates a background distibution based on randomly generated modules of size 'N'.
For each element of the module, pairwise Pearson Correlation (PC), and the Normalized Rank
Euclidean Distance (NRED) are calculated between all genomes. When two genomes contain multiple
elements for comparison, this is handled either by takng a random pairwise comparison,
or by taking the score with the maximum similarity. A background distribution using each method
is provided for comparison. (Currently it only provides the score that minimizes the distances)
}
\examples{
Random_Background_Module_Distances_6<-Background_Distribution_Modules(6,1000)
}
