% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Jaccard.R
\name{Calc_Jaccard}
\alias{Calc_Jaccard}
\title{Calculate Jaccard Distance}
\usage{
Calc_Jaccard(vector1, vector2)
}
\arguments{
\item{vector1, vector2}{Two vectors from which the Jaccard disance is to be calculated}
}
\value{
The Jaccard Distance between two two vectors.
}
\description{
Calculates the Jaccard Distance between two vectors
\deqn{J(A,B) = {{|A \cap B|}\over{|A \cup B|}} = {{|A \cap B|}\over{|A| + |B| - |A \cap B|}}}
}
\examples{
Jaccard<-Calc_Jaccard(vector1,vector2)
}
