% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Z_scores.R
\name{calc_Z_scores}
\alias{calc_Z_scores}
\title{Calcultes mu, sd and Z scores}
\usage{
calc_Z_scores(I_KOs_Background)
}
\arguments{
\item{I_KOs_Background}{a background distribution for KOs}
}
\value{
a list of mu, sd and Z scores of the background distribution
}
\description{
This function calculates the Jaccard Distances between all bins for a given module
}
\examples{
	Z_scores1<-calc_Z_scores(I_KOs_Background)
}
