% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_func.R
\name{cluster_func}
\alias{cluster_func}
\title{Louvain clusters}
\usage{
cluster_func(Zscore_Matrix, Jaccard_Distance)
}
\arguments{
\item{Zscore_Matrix}{An array of all the pairwise Z-scores for a module}

\item{Jaccard_Distance}{An array of all the pairwise Jaccard Distances for a module}
}
\value{
a list containing the average Z score matrix (ave_Z_score_matrix),
the normalized distances after being noralized by Jaccard Distances (JPE_distance),
the JPE distances formatted for a network representation (e.g. cytoscpe),
and the cluster assignments (cl).
}
\description{
This function uses the average composite Z-score normalized by the Jaccard distances to cluster
genome bins to identify clusters with congruent transcriptional responses.
}
\examples{
PHA_clustering_results_P_NRED <-cluster_func(PHA_module_P_NRED$combined,Jaccard_Distance_PHA)
}
