% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterPrune.R
\name{ClusterPrune}
\alias{ClusterPrune}
\title{ClusterPrune}
\usage{
ClusterPrune(clustering_results, Background_Module_Distances)
}
\arguments{
\item{clustering_results}{The results from clustering.}

\item{Background_Module_Distances}{The background distribution for a module of the same size.}
}
\description{
This function calculates p-values for each cluster and returns the pairwise distances in a cluster.
In the future this function will prune the clusters to remove.
}
\examples{
PHA_pvalues_BR<-ClusterPrune(PHA_clustering_results_P_NRED_BR,Random_Background_Module_Distances_6_BR)
}
