% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cor_Matrix.R
\name{Cor_Matrix}
\alias{Cor_Matrix}
\title{Calculates a correlation matrix between two vectors of rows}
\usage{
Cor_Matrix(vector1, vector2, input_matrix, matrix_features)
}
\arguments{
\item{vector1, vector2}{The two lists rows in which a particular KO is identified}

\item{RNAseq_Annotation_Matrix_no_sd_of_zero}{The annotated matrix of RNAseq data with rows that
have a standard deviation of zero removed.}
}
\value{
a pairwise correlation matrix,
}
\description{
This function calculates a pairwise correlation matrix for both pearson and NRED.
Returns an array containing two matrices.
}
\examples{
Cor_Matrix(KO_position_of_A,KO_position_of_B,RNAseq_Annotation_Matrix_no_sd_of_zero)
}
