% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Rank_Columns.R
\name{Create_Rank_Columns}
\alias{Create_Rank_Columns}
\title{Calculate Normalized Ranks (e.g. rank / total genes)}
\usage{
Create_Rank_Columns(RNAseq_Annotated_Matrix)
}
\arguments{
\item{RNAseq_Annotated_Matrix}{The original count matrix (See X for format details).}
}
\value{
The normalized ranks of \code{Sample 1} ... \code{Sample N}.
}
\description{
This function calculates the Normalized Ranks (0-1 scale) of each gene in each genome.
The ranks must be normalized so that they may be compared between genomes of different size.
}
\examples{
RNAseq_Annotated_Matrix<-Create_Rank_Columns(RNAseq_Annotated_Matrix)
}
