% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_ass_matrix.R
\name{fill_association_matrix}
\alias{fill_association_matrix}
\title{Build Association Matrix}
\usage{
fill_association_matrix(clustering_results, matrix_features, module_names)
}
\arguments{
\item{clustering_results}{The output from cluster_func}

\item{matrix_features}{The minimum number of times a KO term must be present to be included in the Matrix}

\item{module_names, }{The names of the modules (ordered)}
}
\value{
A matrix of clustering results. bins are rows and modules are the columns. Values represent the cluster for ecah module
}
\description{
Combine all cluster results into a single matrix
}
\examples{
M_I_association_matrix<-fill_association_matrix(M_I_clustering_results_P_NRED,matrix_features_BR,names(M_I_module_list))
}
