% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Individual_KO_background}
\alias{Individual_KO_background}
\title{Individual_KO_background}
\usage{
Individual_KO_background(RNAseqExpressionCounts, RNAseqExpressionRanks, KOTerms,
  allBins, HighQBins, N)
}
