% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_KOs_Background.R
\name{Individual_KOs_Background}
\alias{Individual_KOs_Background}
\title{Calculating the background distribution for an indiviudal gene}
\usage{
Individual_KOs_Background(RNAseq_Annotation_Matrix_no_sd_of_zero,
  matrix_features, N)
}
\arguments{
\item{RNAseq_Annotation_Matrix_no_sd_of_zero, }{the original matrix with rows that have a standard deviation of zero removed.}

\item{N}{the number of iterations (at least 10000 is suggested)}
}
\value{
A list containting eight vectors of N distances: pearson and euclidean distances for
random, multiple comparison corrected random, random KO, highest scoring pair KO.
}
\description{
This function generates background distributions of the pairwise
similarity in the expression profile of 1) randomly selected genes and 2) randomly
selected genes that share the same function (annontation) in two different genomes.
It is common to have numerous genes with the same annotation in a genome.
This is handled with two approaches, in the first, a single randomly selected pair
of genes with the same function are selected. In the second case, the highest scoring
pair is used. To correct for multiple comparisons when the highest scoring pair is used
the random background distibution is recalculed to include an equal number of multiple
comparisons.
}
\examples{
	Jaccard_Distance_Function(PHA_Module)
}
