% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jaccard_Distance_Function.R
\name{Jaccard_Distance_Function}
\alias{Jaccard_Distance_Function}
\title{Calculte the Jaccard Distance}
\usage{
Jaccard_Distance_Function(Pairwise_Bin_Array_Presence, RNAseq_Annotated_Matrix,
  Module)
}
\value{
A matrix of pairwise Jaccard distnces
}
\description{
This function calculates the Jaccard Distances between all bins for a given module
}
\examples{
	Jaccard_Distance_Function(PHA_Module)
}
