% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_module_lists.R
\name{make_module_lists}
\alias{make_module_lists}
\title{Build module lists}
\usage{
make_module_lists(KO_modules_table)
}
\arguments{
\item{KO_modules_table}{the KEGG database in the structure : WWW, XXX, YYY, ZZZ}
}
\value{
a list of all modules in the KEGG database.
}
\description{
A list of all modules in the KEGG database
}
\examples{
list_of_all_modules <- make_module_lists(KO_modules_table)
}
