% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individual_bd.R
\name{plot_individual_bd}
\alias{plot_individual_bd}
\title{Plot background distribution}
\usage{
plot_individual_bd(I_KOs_Background)
}
\arguments{
\item{I_KOs_Background, }{the background distributions (output from Individual_KOs_Background function)}
}
\value{
TA 4-panel plot
}
\description{
A 4-panel plot
}
\examples{
plot_individual_bd(I_KOs_Background_d_BR)
}
