% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Presence_Absence_Matrix.R
\name{Presence_Absence_Matrix}
\alias{Presence_Absence_Matrix}
\title{Build Presence Absence Matrix of Annotations}
\usage{
Presence_Absence_Matrix(RNAseq_Annotated_Matrix, N)
}
\arguments{
\item{RNAseq_Annotated_Matrix}{The original count matrix (See X for format details).}

\item{N, }{The minimum number of times a KO term must be present to be included in the Matrix}
}
\value{
An X*Y matrix where X = # bins, and Y = # of annotated features.
}
\description{
Using the annotation column(s), this function will build a presence absence matrix
that will be used for calculating the Jaccard Distance. The matrix will be X*Y, where
X is the number of bins, and Y is the number of annotated features.
}
\examples{
Pairwise_Bin_Array_Presence	<- Presence_Absence_Matrix(RNAseq_Annotated_Matrix,5)
}
