% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_func_lists.R
\name{cluster_func}
\alias{cluster_func}
\title{Louvain clustering of Modules}
\usage{
cluster_func(RNAseq_Annotated_Matrix, Composite_Z_Score, matrix_features,
  module_list)
}
\arguments{
\item{RNAseq_Annotated_Matrix}{The original matrix}

\item{Composite_Z_Score}{An array of all the composite Z-scores for each KO.}

\item{module_list}{A list of vectors.}
}
\value{
a list containing the following: a matrix (#bins x #modules in list) storing the cluster information,
a list of Z score matrices for each module (ave_Z_score_matrix),
a list of of matrices with distances after noralization by Jaccard Distances (JPE_distance),
}
\description{
This function identifies clusters of bins with CTR based on the composite Z-scores 
normalized by the Jaccard distances of a cluster.
}
\examples{
M_I_clustering_results_P_NRED <-cluster_func(RNAseq_Annotated_Matrix_BR_default_bin, all_pairwise_KO_distances$Composite_Z_Score, matrix_features_BR, M_I_module_list)
PHA_clustering_results_P_NRED <-cluster_func(RNAseq_Annotated_Matrix_BR_default_bin, all_pairwise_KO_distances$Composite_Z_Score, matrix_features_BR, PHA_module)
}
