% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jaccard_Distance_Function.R
\name{Jaccard_Distance_Function}
\alias{Jaccard_Distance_Function}
\title{Calculte the Jaccard Distance}
\usage{
Jaccard_Distance_Function(RNAseq_Annotated_Matrix, matrix_features, Module)
}
\value{
A matrix of pairwise Jaccard distnces
}
\description{
This function calculates the Jaccard Distances between all bins for a given module
}
\examples{
	J_PHA <-Jaccard_Distance_Function(RNAseq_Annotated_Matrix_BR_default_bin, matrix_features_BR, PHA_module)
}
